Upgrade to rel/7.0.0

+++ UnitTests/UnitTests.HighPerformance.Shared/Extensions/Test_BoolExtensions.cs
@@ -3,7 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using Microsoft.Toolkit.HighPerformance.Extensions;
+using Microsoft.Toolkit.HighPerformance;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace UnitTests.HighPerformance.Extensions
@@ -15,16 +15,18 @@ namespace UnitTests.HighPerformance.Extensions
         [TestMethod]
         public void Test_BoolExtensions_True()
         {
-            Assert.AreEqual(1, true.ToInt(), nameof(Test_BoolExtensions_True));
-            Assert.AreEqual(1, (DateTime.Now.Year > 0).ToInt(), nameof(Test_BoolExtensions_True));
+            // There tests all just run a couple of boolean expressions and validate that the extension
+            // correctly produces either 1 or 0 depending on whether the expression was true or false.
+            Assert.AreEqual(1, true.ToByte(), nameof(Test_BoolExtensions_True));
+            Assert.AreEqual(1, (DateTime.Now.Year > 0).ToByte(), nameof(Test_BoolExtensions_True));
         }
 
         [TestCategory("BoolExtensions")]
         [TestMethod]
         public void Test_BoolExtensions_False()
         {
-            Assert.AreEqual(0, false.ToInt(), nameof(Test_BoolExtensions_False));
-            Assert.AreEqual(0, (DateTime.Now.Year > 3000).ToInt(), nameof(Test_BoolExtensions_False));
+            Assert.AreEqual(0, false.ToByte(), nameof(Test_BoolExtensions_False));
+            Assert.AreEqual(0, (DateTime.Now.Year > 3000).ToByte(), nameof(Test_BoolExtensions_False));
         }
 
         [TestCategory("BoolExtensions")]
